requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dnspython==1.16.0
filelock==3.16.1
flask==3.0.3
flask-cors==5.0.0
flask-mail==0.10.0
flatbuffers==24.3.25
gast==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.152.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.66.2
grpcio-status==1.62.3
h5py==3.12.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
libclang==18.1.1
lxml==5.3.0
markdown==3.7
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pyjwt==2.9.0
pymongo==3.12.0
pyparsing==3.2.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.1
rsa==4.9
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.5.0
tqdm==4.67.0
tweepy==4.14.0
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
